+++ packages/mangrove-solidity/test/lib/Test2.sol
@@ -58,6 +58,14 @@ contract Test2 is Test, Utilities {
     }
   }
 
+  /* *** Test & cheatcode helpers **** */
+
+  modifier prank(address a) {
+    vm.startPrank(a);
+    _;
+    vm.stopPrank();
+  }
+
   /* sugar for successful test */
   function succeed() internal {
     assertTrue(true);
@@ -82,34 +90,33 @@ contract Test2 is Test, Utilities {
      does not have to be unique.
 
      Variants:
-     freshAccount()     :         key,address
-     freshAccount(label): labeled key,address
-     freshKey()         :         key
-     freshAddress()     :         address
-     freshAddress(label): labeled address
+     freshAccount(label)       labeled key,address
+     freshAccount()            key,address
+     freshKey()                key
+     freshAddress(label)       labeled address
+     freshAddress()            address
   */
-  function freshAccount() internal returns (uint, address payable) {
-    uint key = keyIterator++;
-    address payable addr = payable(vm.addr(key));
-    // set code to nonzero so solidity-inserted extcodesize checks don't fail
-    vm.etch(addr, bytes("not zero"));
-    return (key, addr);
-  }
-
   function freshAccount(string memory label)
     internal
     returns (uint key, address payable addr)
   {
-    (key, addr) = freshAccount();
+    unchecked {
+      key = (keyIterator++) + uint(keccak256(bytes(label)));
+    }
+    addr = payable(vm.addr(key));
+
+    // set code to nonzero so solidity-inserted extcodesize checks don't fail
+    vm.etch(addr, bytes("not zero"));
     vm.label(addr, label);
+    return (key, addr);
+  }
+
+  function freshAccount() internal returns (uint, address payable) {
+    return freshAccount(vm.toString(keyIterator));
   }
 
   function freshKey() internal returns (uint key) {
     (key, ) = freshAccount();
-  }
-
-  function freshAddress() internal returns (address payable addr) {
-    (, addr) = freshAccount();
   }
 
   function freshAddress(string memory label)
@@ -119,22 +126,17 @@ contract Test2 is Test, Utilities {
     (, addr) = freshAccount(label);
   }
 
+  function freshAddress() internal returns (address payable addr) {
+    (, addr) = freshAccount();
+  }
+
   /* expect exact log from address */
   function expectFrom(address addr) internal {
     vm.expectEmit(true, true, true, true, addr);
   }
 
-  /* expect a revert reason */
-  function revertEq(string memory actual_reason, string memory expected_reason)
-    internal
-    returns (bool)
-  {
-    assertEq(actual_reason, expected_reason, "wrong revert reason");
-    return true;
-  }
-
   /* assert address is not 0 */
-  function not0x(address a) internal returns (bool) {
+  function assertNot0x(address a) internal returns (bool) {
     if (a == address(0)) {
       emit log("Error: address should not be 0");
       emit log_named_address("Address", a);
@@ -143,7 +145,8 @@ contract Test2 is Test, Utilities {
     return (a != address(0));
   }
 
-  /* inline gas measures */
+  /* *** Gas Metering *** */
+
   uint private checkpointGasLeft = 1; // Start the slot non0.
   string checkpointLabel;
 
@@ -162,29 +165,5 @@ contract Test2 is Test, Utilities {
     uint gasDelta = checkpointGasLeft - checkpointGasLeft2 - 100;
 
     emit log_named_uint("Gas used", gasDelta);
-  }
-
-  /* Logging is put here since solidity libraries cannot be extended. */
-
-  function logary(uint[] memory uints) public view {
-    string memory s = "";
-    for (uint i = 0; i < uints.length; i++) {
-      s = string.concat(s, vm.toString(uints[i]));
-      if (i < uints.length - 1) {
-        s = string.concat(s, ", ");
-      }
-    }
-    console2.log(s);
-  }
-
-  function logary(int[] memory ints) public view {
-    string memory s = "";
-    for (uint i = 0; i < ints.length; i++) {
-      s = string.concat(s, vm.toString(uint(ints[i])));
-      if (i < ints.length - 1) {
-        s = string.concat(s, ", ");
-      }
-    }
-    console2.log(s);
   }
 }