+++ packages/mangrove-solidity/test/lib/forks/Generic.sol
@@ -1,13 +1,12 @@
 // SPDX-License-Identifier:	AGPL-3.0
 pragma solidity ^0.8.10;
 
-import "mgv_test/lib/MangroveTest.sol";
-import "forge-std/Vm.sol";
+import {Script} from "forge-std/Script.sol";
 
 contract GenericFork is Script {
   uint public INTERNAL_FORK_ID;
   uint public CHAIN_ID;
-  string public NAME;
+  string public NAME = "generic";
   uint public BLOCK_NUMBER;
 
   address public AAVE;
@@ -22,18 +21,36 @@ contract GenericFork is Script {
   address public CUSDC;
   address public CWETH;
 
+  function roll() public {
+    vm.rollFork(INTERNAL_FORK_ID);
+  }
+
+  function roll(uint blockNumber) public {
+    vm.rollFork(INTERNAL_FORK_ID, blockNumber);
+  }
+
+  function select() public {
+    vm.selectFork(INTERNAL_FORK_ID);
+  }
+
   function setUp() public virtual {
-    vm.label(AAVE, "Aave");
-    vm.label(APOOL, "Aave Pool");
-    vm.label(WETH, "WETH");
-    vm.label(AUSDC, "AUSDC");
-    vm.label(USDC, "USDC");
-    vm.label(AWETH, "AWETH");
-    vm.label(DAI, "DAI");
-    vm.label(ADAI, "ADAI");
-    vm.label(CDAI, "CDAI");
-    vm.label(CUSDC, "CUSDC");
-    vm.label(CWETH, "CWETH");
+    if (CHAIN_ID == 0) {
+      revert(
+        "No fork selected: you should pick a subclass of GenericFork with a nonzero CHAIN_ID."
+      );
+    }
+
+    label(AAVE, "Aave");
+    label(APOOL, "Aave Pool");
+    label(WETH, "WETH");
+    label(AUSDC, "AUSDC");
+    label(USDC, "USDC");
+    label(AWETH, "AWETH");
+    label(DAI, "DAI");
+    label(ADAI, "ADAI");
+    label(CDAI, "CDAI");
+    label(CUSDC, "CUSDC");
+    label(CWETH, "CWETH");
 
     vm.makePersistent(address(this));
 
@@ -45,12 +62,6 @@ contract GenericFork is Script {
     }
 
     vm.selectFork(INTERNAL_FORK_ID);
-
-    if (CHAIN_ID == 0) {
-      revert(
-        "No fork selected: you should pick a subclass of GenericFork with a nonzero CHAIN_ID."
-      );
-    }
 
     if (block.chainid != CHAIN_ID) {
       revert(
@@ -64,5 +75,9 @@ contract GenericFork is Script {
         )
       );
     }
+  }
+
+  function label(address addr, string memory str) internal {
+    vm.label(addr, string.concat(str, "(", NAME, ")"));
   }
 }